app/requirements.txt
# Django
Django>=3.0,<4.0

# Base
psycopg2-binary>=2.8

# Django REST framework
djangorestframework==3.12.2
django-filter==2.3.0

# Environment
requests==2.25.1
django-environ==0.4.5
